feature/dep315: Fixed api linting errors, fixed failing web unit test.
