+++ web/src/components/landing/utils.ts
@@ -1,4 +1,4 @@
-import { tryParse } from 'engagements/admin/create/authoring/utils';
+import { tryParse } from 'helper';
 import { SearchFilters } from './types';
 
 export const updateSearchParams = (newFilters: SearchFilters, searchParams: URLSearchParams) => {
